% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{Agronomic_features}
\alias{Agronomic_features}
\title{Agronomic_features}
\format{A list of Agronomic Features}
\source{
CGIAR Big Data Platform, Bioversity, International Potato Center, CIAT,  CGIAR
}
\usage{
Agronomic_features
}
\description{
This list contains the list of Agronomic Features. Users can pick up elements on this list trough ShinyTree
}
\references{
This data is related to HiDAP crop template
}
