% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{agronomic_trait_list}
\alias{agronomic_trait_list}
\title{agronomic_trait_list}
\format{A list of Agronomic Traits}
\source{
CGIAR Big Data Platform, Bioversity, International Potato Center, CIAT,  CGIAR
}
\usage{
agronomic_trait_list
}
\description{
This list contains the list of Agronomic Traits. Users can pick up elements on this list trough ShinyTree
}
\references{
This data is related to HiDAP crop template
}
