% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{big_crop_template_xlsx}
\alias{big_crop_template_xlsx}
\title{Special or Big Crop Template in HiDAP}
\format{A data frame data.frame which comes from importing excel crop template files.}
\source{
International Potato Center, potato experimental data.
}
\usage{
big_crop_template_xlsx
}
\description{
This dataset contains big crop template to create special or big modules using rda format.
}
\references{
This data is related to HiDAP big or special crop template
}
