% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_values.R
\name{big_get_tree_value}
\alias{big_get_tree_value}
\title{Get values from shinyTree multiple objects}
\usage{
big_get_tree_value(tree_input_value, crop_selected)
}
\arguments{
\item{tree_input_value}{Input values selected from shinyTree objects}

\item{crop_selected}{the name of the crop}
}
\description{
This function gets all the selected values from shinyTree checkboxes in Big Modules or Special Modules
}
