% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_type_list_fname}
\alias{get_type_list_fname}
\alias{get_mlist_file_name}
\title{Get the type of list (clonal, family or parental list) according to the file name.}
\usage{
get_type_list_fname(mlist)

get_mlist_file_name(type_import, ml_file_name)
}
\arguments{
\item{mlist}{SelectInput value (commonly),}

\item{type_import}{shiny input value. SelectInput value for type of import}

\item{ml_file_name}{shiny input value. SelecInput value for file names}
}
\description{
Get the type of list (clonal, family or parental list) according to the file name.

Detection of parent list
}
\section{Functions}{
\itemize{
\item \code{get_type_list_fname}: Say \code{parent} wheter it is a parental list. Otherwise, \code{standard} whether it's a clonal or family list.

\item \code{get_mlist_file_name}: Say \code{parent} wheter it is a parental list. Otherwise, \code{standard} whether it's a clonal or family list.
}}

\author{
Omar Benites

Omar Benites
}
