% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{crop_list}
\alias{crop_list}
\title{List of potato and sweetpotato tratis in HiDAP}
\format{A list with potato and sweetpotato traits}
\source{
International Potato Center, potato andsweetpotato experimental data.
}
\usage{
crop_list
}
\description{
This dataset contains the potato and sweetpotato trait list
}
\references{
This data is related to HiDAP crop template
}
