% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{crop_template_xlsx}
\alias{crop_template_xlsx}
\title{Crop Template for Excel Fieldbooks in HiDAP}
\format{A data frame data.frame which comes from importing excel crop template files.}
\source{
International Potato Center, potato experimental data.
}
\usage{
crop_template_xlsx
}
\description{
This dataset contains the excel crop template to create Fieldbooks
}
\references{
This data is related to HiDAP crop template
}
