% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_design.R
\name{design.alpha.check}
\alias{design.alpha.check}
\title{Alpha Design Condition Checker}
\usage{
design.alpha.check(trt, k, r)
}
\arguments{
\item{trt}{Treatments}

\item{k}{Block size}

\item{r}{Replication}
}
\description{
Function to check the necessary conditions
}
