% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_design.R
\name{design.f2rcbd}
\alias{design.f2rcbd}
\title{Factorial Two-Way Design under Complete Block Randomization}
\usage{
design.f2rcbd(trt1, trt2, r, series = 1, random = TRUE)
}
\arguments{
\item{trt1}{Treatments}

\item{trt2}{The factorial level}

\item{r}{Replications or blocks}

\item{series}{The plot start number.  1: 11,12; 2: 101,102; 3: 1001,1002}

\item{random}{TRUE or FALSE - randomize}
}
\description{
Creates an factorial two-Way design under Complete Block Randomization
}
\author{
Omar Benites
}
