% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_design.R
\name{design.undr}
\alias{design.undr}
\title{Unreplicated design with no-randomization}
\usage{
design.undr(trt, r = 1)
}
\arguments{
\item{trt}{Treatments}

\item{r}{A Pseudo replications. It's just preserve the order. By default r=1.}
}
\description{
Creates an unreplicated design with no-randomization.
}
\author{
Omar Benites
}
