% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{dt_cmea}
\alias{dt_cmea}
\title{dt_cmea Crop measurement dataset related to Agronomic Trials}
\format{data frame}
\source{
CGIAR Big Data Platform, Bioversity, International Potato Center CIP, CIAT,  CGIAR
}
\usage{
dt_cmea
}
\description{
dataset related to crop measurement
}
\references{
This data is related to HiDAP-Agromfims Platform
}
