% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{dt_factordesign}
\alias{dt_factordesign}
\title{dt_factordesign Data for Factor Variables}
\format{data.frame}
\source{
CGIAR Big Data Platform, Bioversity, International Potato Center, CIAT,  CGIAR
}
\usage{
dt_factordesign
}
\description{
Table of factor and levels for the design tab
}
\references{
This data is related to AGROFIMS crop template
}
