% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{dt_fernut}
\alias{dt_fernut}
\title{dt_fernut}
\format{List/Map}
\source{
CGIAR Big Data Platform, Bioversity, International Potato Center, CIAT,  CGIAR
}
\usage{
dt_fernut
}
\description{
fertilizer and nutrient dataset
}
\references{
A set of data related to AGROFIMS Fertility Module
}
