% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_inputs}
\alias{dt_inputs}
\title{Add columns to data frames with different sizes}
\usage{
dt_inputs(dt, dt_other)
}
\arguments{
\item{dt}{data frame}

\item{dt_other}{vector New column to add into the data frame}
}
\description{
Add columns to data frames with different sizes
}
\author{
Omar Benites
}
