% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{Evaluation_Forms}
\alias{Evaluation_Forms}
\title{A list with evaluation and forms used in Participatory Varietal Selection}
\format{A list used in potato and sweetpotato traits}
\source{
International Potato Center, potato andsweetpotato experimental data.
}
\usage{
Evaluation_Forms
}
\description{
This list contains all forms and trait used in Participatory Varietal Selection. Users can pick up elements on this list trough ShinyTree
}
\references{
This data is related to HiDAP crop template
}
