% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{extra_variables}
\alias{extra_variables}
\title{extra_variables}
\format{data.frame}
\source{
CGIAR Big Data Platform, Bioversity, International Potato Center, CIAT,  CGIAR
}
\usage{
extra_variables
}
\description{
A list of extra variables that should be placed in TraitList for KDSMART
}
\references{
This data is related to AGROFIMS crop template
}
