% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-design-fun.R
\name{fctinputs}
\alias{fctinputs}
\title{Get levels of factors (inputs) inserted by users in AGROFIMS}
\usage{
fctinputs(gr1 = NULL, gr2 = NULL, gr3 = NULL, napp = 3, rtf, ref,
  dfr)
}
\arguments{
\item{gr1}{group 1 for selected factor in UI}

\item{gr2}{group 2 for selected factor in UI}

\item{gr3}{group 3 for selected factor in UI}

\item{napp}{numeric Number of application}

\item{rtf}{numeric Rate total producto}

\item{ref}{numeric Nutrient rate element}

\item{dfr}{data frame Selected factors builded on a data frame}
}
\description{
Get inputs for experimental design tab
}
\author{
Omar Benites
}
