% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{fertCombo}
\alias{fertCombo}
\title{fertCombo}
\format{List/Map}
\source{
CGIAR Big Data Platform, Bioversity, International Potato Center, CIAT,  CGIAR
}
\usage{
fertCombo
}
\description{
fertilizer technique, timing and implement combo values in AGROFIMS
}
\references{
This data is related to AGROFIMS Fertility Module
}
