% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geo_template.R
\docType{data}
\name{geodb}
\alias{geodb}
\title{geodb}
\format{A list of countries data in HiDAP-AGROFIMS}
\source{
CGIAR Big Data Platform, Bioversity, International Potato Center CIP, CIAT,  CGIAR
}
\usage{
geodb
}
\description{
This list contains the list of countries and localities over the globe. Users can pick up elements on this list trough selectInput.
}
\references{
This data is related to HiDAP-Agromfims crop template
}
\author{
Robert Hijmans, UC Davis
}
