% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_resmgt.R
\name{get_ec_resmgt}
\alias{get_ec_resmgt}
\title{Get residual management inputs}
\usage{
get_ec_resmgt(input, lbl)
}
\arguments{
\item{input}{shiny reactiveValuesList provided by AGROFIMS}

\item{lbl}{\code{character} vector labels}
}
\description{
get all input values related to the experiment conditions \code{residue management} experiment condition tab, 
provided by users when type in AGROFIMS.
}
