% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_soilfertility.R
\name{get_ec_sf}
\alias{get_ec_sf}
\title{Get exp conditions , soil and fertilkity tab}
\usage{
get_ec_sf(allinputs, napp)
}
\arguments{
\item{allinputs}{shiny reactiveValuesList provided by AGROFIMS}

\item{napp}{\code{numeric} number of applications}
}
\description{
get all input values related to the experiment conditions \code{soil and fertility} experiment condition tab, 
provided by users when type in AGROFIMS.
}
\author{
Omar Benites
}
