% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_soilfertility.R
\name{get_fertilizer}
\alias{get_fertilizer}
\title{Get fertilizer product names from Fertilizer and Nutrient dataset}
\usage{
get_fertilizer(.data)
}
\arguments{
\item{.data}{Fertilizer and Nutrient dataset}
}
\description{
Get fertilizer's names from Fertilizer and Nutrient dataset
}
\author{
Omar Benites
}
