% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_design.R
\name{get_fertilizer_details_magm}
\alias{get_fertilizer_details_magm}
\title{TODO Product Amount}
\usage{
get_fertilizer_details_magm(allinputs, indexSoilMagp, indexProdSplit)
}
\description{
TODO Product Amount
}
