% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_loop_AgrOper}
\alias{get_loop_AgrOper}
\title{Get agronomic operations inputs from experiment conditions}
\usage{
get_loop_AgrOper(feature = "", n, label = "none", other = "other")
}
\arguments{
\item{feature}{value of the feature}

\item{n}{number of values}

\item{label}{an argument to get units from field operations}

\item{other}{character. Specify 'other' value inputs.}
}
\description{
Insert column between columns based on positions.
}
