% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_loop_irrigation_technique}
\alias{get_loop_irrigation_technique}
\title{Get techniques inputs}
\usage{
get_loop_irrigation_technique(technique, module = "irrigation",
  submodule = "_type_")
}
\arguments{
\item{technique}{vector. Vector with technique values.}

\item{module}{character. Name of the module.}

\item{submodule}{character. Name of the submodule with underscore symbol.}
}
\description{
Get techniques inputs
}
