% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_loop_irrigation_technique_other}
\alias{get_loop_irrigation_technique_other}
\title{Get inputs according to techniques and subtechniques inputs}
\usage{
get_loop_irrigation_technique_other(technique = "Irrigation sprinker",
  subtechnique = "Other")
}
\arguments{
\item{technique}{vector. Vector with technique values.}

\item{subtechnique}{vector.Vector with sub technique values.}
}
\description{
Get inputs according to techniques and subtechniques inputs
}
