% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_design.R
\name{get_nutrient_details_magm}
\alias{get_nutrient_details_magm}
\title{TODO: Obtener las tablas "sfNutrientSplit_", con los valores de cada elemento}
\usage{
get_nutrient_details_magm(allinputs, addId)
}
\description{
TODO: Obtener las tablas "sfNutrientSplit_", con los valores de cada elemento
}
