% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_type_list_ds}
\alias{get_type_list_ds}
\title{Type of material list (clonal, family or parental list) according to the data structure.}
\usage{
get_type_list_ds(mlist)
}
\arguments{
\item{mlist}{list. List of attributes based on breeding material tables.}
}
\description{
Logical. Say TRUE if your material belongs to parental list.
}
\author{
Omar Benites
}
