% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAddInputId}
\alias{getAddInputId}
\title{Get input id derived from AddButton event}
\usage{
getAddInputId(addId = "", pattern = "FA_", replacement = "")
}
\arguments{
\item{addId}{character vector input id(s)}

\item{pattern}{character pattern to replace}

\item{replacement}{character value to replace some pattern}
}
\description{
Get input id derived from AddButton event
}
\author{
Omar Benites
}
