% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAgrOper}
\alias{getAgrOper}
\title{Get agronomic operations inputs from experiment conditions}
\usage{
getAgrOper(feature, other = "")
}
\arguments{
\item{feature}{value of the feature}

\item{other}{other value}
}
\description{
Insert column between columns based on positions.
}
