% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTrtInputs}
\alias{getTrtInputs}
\title{Get treatment and factor inputs from design of experiments}
\usage{
getTrtInputs(group, subgroup, fct, dfr)
}
\arguments{
\item{group}{group}

\item{subgroup}{subgroup}

\item{fct}{factor}

\item{dfr}{data frame with the inputs}
}
\description{
Insert column between columns based on positions.
}
