% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabel.R
\name{h_experiment_details}
\alias{h_experiment_details}
\title{Header's vector for experiment detail placed on Experiment tab}
\usage{
h_experiment_details()
}
\description{
Header's vector for experiment detail placed on Experiment tab
}
