% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_parentList}
\alias{is_parentList}
\title{Detection of parent list}
\usage{
is_parentList(mlist_name)
}
\arguments{
\item{mlist_name}{SelectInput value (commonly),}
}
\description{
Detection of parent list
}
\section{Functions}{
\itemize{
\item \code{is_parentList}: Say TRUE if your material belongs to parental list.
}}

\author{
Omar Benites
}
