% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{magm_label}
\alias{magm_label}
\title{magm_label}
\format{data.frame}
\source{
CGIAR Big Data Platform, Bioversity, International Potato Center, CIAT,  CGIAR
}
\usage{
magm_label
}
\description{
Labels for management practices UI in AGROFIMS
}
\references{
This data is related to AGROFIMS Management Practices
}
