% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{magmtprac}
\alias{magmtprac}
\title{magmtprac}
\format{data.frame}
\source{
CGIAR Big Data Platform, Bioversity, International Potato Center, CIAT,  CGIAR
}
\usage{
magmtprac
}
\description{
Table of Management Practices
}
\references{
This data is related to AGROFIMS Management Practices
}
