% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-temp.R
\name{map_level_values}
\alias{map_level_values}
\title{Mimic the functionality of map functions and transform level inputs into data frame}
\usage{
map_level_values(input, allinputs, isf = c("yes", "no"), id_type_dt,
  id_rand, ntrt = 2, lbl = "f")
}
\arguments{
\item{input}{shinyInput input values from server side}

\item{allinputs}{reactive shiny object that return a data frame with all the shiny inputs}

\item{isf}{character If \code{isf=="yes"} is full factorial, otherwise \code{isf=="no"} is for non-full factorial}

\item{id_type_dt}{data.frame Table with type of Factor and types of input form: \code{combo box}, \code{text input}, and \code{date}}

\item{id_rand}{character vector character random id}

\item{ntrt}{If \code{isf=="no"}, introduce the number of treatments in non-full factorials arragements.}

\item{lbl}{characer Optional argument, in case format = "data.frame"}
}
\description{
Mimic the functionality of map functions and transform level inputs into data frame
}
