% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_values}
\alias{map_values}
\title{Mimic the functionality of map functions and transform inputs by applying a function to each element and returning a vector the same length as the input.}
\usage{
map_values(input, id_chr = "", id_rand, format = c("vector", "data.frame"),
  lbl = NULL)
}
\arguments{
\item{input}{shinyInput input values from server side}

\item{id_chr}{character character pattern id}

\item{id_rand}{character vector character random id}

\item{format}{character object format to be exported. "vector" for vector data structure or "data.frame" for data frame structure}

\item{lbl}{characer Optional argument, in case format = "data.frame"}
}
\description{
Mimic the functionality of map functions and transform inputs by applying a function to each element and returning a vector the same length as the input.
}
