% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{material_list}
\alias{material_list}
\title{A data frame with material list headers}
\format{A a data frame used in potato and sweetpotato traits}
\source{
International Potato Center, potato andsweetpotato experimental data.
}
\usage{
material_list
}
\description{
This data frame contains all the compulsory header used in Material List excel files for
}
\references{
This data is related to HiDAP crop template
}
