% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{potato_list}
\alias{potato_list}
\title{List of potato tratis in HiDAP}
\format{A list with potato traits}
\source{
International Potato Center, potato experimental data.
}
\usage{
potato_list
}
\description{
This dataset contains the potato trait list
}
\references{
This data is related to HiDAP crop template
}
