% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{Selection_Criteria}
\alias{Selection_Criteria}
\title{Selection_Criteria}
\format{A list used selection criteria}
\source{
International Potato Center, potato and sweetpotato experimental data.
}
\usage{
Selection_Criteria
}
\description{
This list contains a selection criterias list for Participatory Varietal Selection. Users can pick up elements on this list trough ShinyTree
}
\references{
This data is related to HiDAP crop template
}
