% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_listsites_agrofims.R
\name{server_listsites_agrofims}
\alias{server_listsites_agrofims}
\title{server_design for managing server functionalities of list sites on HIDAP-AGROFIMS}
\usage{
server_listsites_agrofims(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Design a fieldbook for HIDAP-AGROFIMS
}
\author{
Omar Benites/Ivan Perez/Raul Arias
}
