% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_mobile_agrofims.R
\name{server_mobile_agrofims}
\alias{server_mobile_agrofims}
\title{shiny UI element for HIDAP-KDSMART}
\usage{
server_mobile_agrofims(input, output, session, values)
}
\arguments{
\item{type}{type of ui element; default is a tab in a shiny dashboard}

\item{title}{display title}

\item{name}{a reference name}
}
\description{
returns a re-usable user interface element HIDAP-KDSMART
}
\author{
Ivan Perez, Omar Benites
}
