% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{setUserSession}
\alias{setUserSession}
\title{set user session}
\usage{
setUserSession(logged, username, id, name = NULL, lastname = NULL)
}
\arguments{
\item{logged}{Bool user is logged}

\item{username}{user username}

\item{id}{user id}

\item{name}{user name}

\item{lastname}{user lastname}
}
\description{
set user session
}
\author{
Raul Arias
}
