% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{site_table}
\alias{site_table}
\title{List of trial sites}
\format{A list with trial location sites}
\source{
International Potato Center, potato andsweetpotato experimental data.
}
\usage{
site_table
}
\description{
This dataset contains the master trial sites
}
\references{
This data is related to HiDAP trial sites
}
