% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{soil_list}
\alias{soil_list}
\title{soil_list}
\format{A list of Soil variables}
\source{
CGIAR Big Data Platform, Bioversity, International Potato Center, CIAT,  CGIAR
}
\usage{
soil_list
}
\description{
This list contains the list of Soil variables. Users can pick up elements on this list trough ShinyTree
}
\references{
This data is related to HiDAP crop template
}
