% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{sweetpotato_list}
\alias{sweetpotato_list}
\title{List of sweetpotato tratis in HiDAP}
\format{A list with sweetpotato traits}
\source{
International Potato Center, sweetpotato experimental data.
}
\usage{
sweetpotato_list
}
\description{
This dataset contains the sweetpotato trait list
}
\references{
This data is related to HiDAP crop template
}
