% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{table_crops}
\alias{table_crops}
\title{List of crops in HiDAP}
\format{A data frame data.frame with potato and sweetpotato crops description.}
\source{
International Potato Center, potato experimental data.
}
\usage{
table_crops
}
\description{
This dataset contains the crops used in HiDAP
}
\references{
This data is related to HiDAP crop template
}
