% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{table_module_big}
\alias{table_module_big}
\title{table of special or big modules}
\format{A data frame data.frame which comes from importing excel potato and sweetpotato short table.}
\source{
International Potato Center, potato experimental data.
}
\usage{
table_module_big
}
\description{
This dataset contains the trait used in Potato and Sweetpotato Sepcial Trials
}
\references{
This data is related to HiDAP special or big crop_template.
}
