% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{table_module_potato}
\alias{table_module_potato}
\title{Potato traits}
\format{A data frame data.frame which comes from importing excel potato dictionary file.}
\source{
International Potato Center, potato experimental data.
}
\usage{
table_module_potato
}
\description{
This dataset contains the trait used in Potato Trials
}
\references{
This data is related to HiDAP crop template
}
