% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{table_module_sweetpotato}
\alias{table_module_sweetpotato}
\title{Sweet Potato Traits}
\format{A data frame data.frame which comes from importing excel sweetpotato dictionary file.}
\source{
International Potato Center, sweetpotato experimental data.
}
\usage{
table_module_sweetpotato
}
\description{
This dataset contains the trait used in SweetPotato Trials
}
\references{
This data is related to HiDAP crop template
}
