% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{Trait_List}
\alias{Trait_List}
\title{List of potato and sweetpotato tratis in HiDAP}
\format{A list with potato and sweetpotato traits dictionary}
\source{
International Potato Center, potato andsweetpotato experimental data.
}
\usage{
Trait_List
}
\description{
This crop list contains the potato and sweetpotato trait list used in shinyTree
}
\references{
This data is related to HiDAP crop template
}
