% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_design_hdagrofims.R
\name{ui_fieldbook_agrofims}
\alias{ui_fieldbook_agrofims}
\title{shiny UI element for HIDAP-AGROFIMS}
\usage{
ui_fieldbook_agrofims(type = "tab", title = "Design Fieldbook",
  name = "phenotype_fieldbook_design")
}
\arguments{
\item{type}{type of ui element; default is a tab in a shiny dashboard}

\item{title}{display title}

\item{name}{a reference name}
}
\description{
returns a re-usable user interface element
}
\author{
Raul Arias, Omar Benites, Ivan Perez
}
