% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_listsites_agrofims.R
\name{ui_listsites_agrofims}
\alias{ui_listsites_agrofims}
\title{UI List Sites for managing new trial site in for HIDAP-AGROFIMS
Returns user friendly ui for for HIDAP-AGROFIMS}
\usage{
ui_listsites_agrofims(type = "tab", title = "", name = "listSites")
}
\arguments{
\item{type}{type of UI element, default is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
UI List Sites for managing new trial site in for HIDAP-AGROFIMS
Returns user friendly ui for for HIDAP-AGROFIMS
}
\author{
Omar Benites/Ivan Perez/Raul Arias
}
