% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_mobile_agrofims.R
\name{ui_mobile_agrofims}
\alias{ui_mobile_agrofims}
\title{shiny UI element for HIDAP-KDSMART}
\usage{
ui_mobile_agrofims(type = "tab", title = "KDSmart manage",
  name = "kdsmart_fieldbook_design")
}
\arguments{
\item{type}{type of ui element; default is a tab in a shiny dashboard}

\item{title}{display title}

\item{name}{a reference name}
}
\description{
returns a re-usable user interface element HIDAP-KDSMART
}
\author{
Ivan Perez, Omar Benites
}
