% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_session_agrofims.R
\name{ui_session_agrofims}
\alias{ui_session_agrofims}
\title{shiny UI element for HIDAP-SESSION}
\usage{
ui_session_agrofims(type = "tab", title = "Session manage",
  name = "session_fieldbook_design")
}
\arguments{
\item{type}{type of ui element; default is a tab in a shiny dashboard}

\item{title}{display title}

\item{name}{a reference name}
}
\description{
returns a re-usable user interface element HIDAP-SESSION
}
\author{
Ivan Perez, Omar Benites
}
