% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_sites_agrofims.R
\name{ui_sites_agrofims}
\alias{ui_sites_agrofims}
\title{UI Side for managing new trial site in for HIDAP-AGROFIMS
Returns user friendly ui for for HIDAP-AGROFIMS}
\usage{
ui_sites_agrofims(type = "tab", title = "New Trial Site",
  name = "createList")
}
\arguments{
\item{type}{type of UI element, default is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
UI Side for managing new trial site in for HIDAP-AGROFIMS
Returns user friendly ui for for HIDAP-AGROFIMS
}
\author{
Omar Benites
}
