% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{weather_list}
\alias{weather_list}
\title{weather_list}
\format{A list of Weather variables}
\source{
CGIAR Big Data Platform, Bioversity, International Potato Center, CIAT,  CGIAR
}
\usage{
weather_list
}
\description{
This list contains the list of Weather variables. Users can pick up elements on this list trough ShinyTree
}
\references{
This data is related to HiDAP crop template
}
