% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crop_template.R
\docType{data}
\name{weather_station_vars}
\alias{weather_station_vars}
\title{weather_station_vars weather soil_data}
\format{A list of installation sheet data in HiDAP-AGROFIMS}
\source{
CGIAR Big Data Platform, Bioversity, International Potato Center CIP, CIAT,  CGIAR
}
\usage{
weather_station_vars
}
\description{
This list contains the list of installation sheet data. Users can pick up elements on this list trough ShinyTree
}
\references{
This data is related to HiDAP-Agromfims crop template
}
