% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add.col}
\alias{add.col}
\title{Add columns to data frames with different sizes}
\usage{
add.col(df, new.col)
}
\arguments{
\item{df}{data frame}

\item{new.col}{vector New column to add into the data frame}
}
\description{
Add columns to data frames with different sizes
}
\references{
https://stat.ethz.ch/pipermail/r-help/2004-October/059752.html
}
