% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_desing_hdagrofim.R
\name{server_design_agrofims}
\alias{server_design_agrofims}
\title{Design field book for HIDAP-AGROFIMS}
\usage{
server_design_agrofims(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Design field book for HIDAP-AGROFIMS
}
\author{
Omar Benites
}
